+++ lib/core.dart
@@ -5,4 +5,5 @@ library;
 export 'src/core/dimensions.dart';
 export 'src/core/exception.dart';
 export 'src/core/quantity.dart';
+export 'src/core/scalar.dart';
 export 'src/core/unit.dart';