+++ lib/units.dart
@@ -4,8 +4,14 @@
 /// Can pollute the global namespace, so only import if required, or use a prefix on import.
 library;
 
-export 'src/common/units.dart';
-export 'system/imperial.dart';
-export 'system/international.dart';
-export 'system/si.dart';
-export 'system/usc.dart';
+import 'units/imperial.dart' as imp show bushel;
+import 'units/usc.dart' as usc show bushel;
+
+export 'units/imperial.dart' hide bushel;
+export 'units/international.dart';
+export 'units/scalar.dart';
+export 'units/si.dart';
+export 'units/usc.dart' hide bushel;
+
+const bushelImp = imp.bushel;
+const bushelUs = usc.bushel;