+++ lib/units/international.dart
@@ -1,6 +1,6 @@
-/// A library containing imperial [Unit] definitions.
+/// A library containing international [Unit] definitions.
 ///
 /// Can pollute the global namespace, so only import if required, or use a prefix on import.
 library;
 
-export '../src/system/imperial.dart';
+export '../src/units/international.dart';