+++ lib/units/scalar.dart
@@ -1,0 +1,6 @@
+/// A library containing scalar [Unit] definitions.
+///
+/// Can pollute the global namespace, so only import if required, or use a prefix on import.
+library;
+
+export '../src/units/scalar.dart';