+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 0.6.0
+  * Renamed `systems` to `units`
+  * Moved scalar units to `units/scalar.dart`
+  * Quantity.scaled() now accepts a `num` as well as a `Quantity`
+
 ### 0.5.1
   * Fixed prefixed unit scaling [2c393b1](https://github.com/nebkat/dart-physical/commit/2c393b1a9836ca5ea5aae435c866167feceda6b4)
   * Reorganized examples